requirements.txt
Flask==3.0.2
pytest==8.0.0
pytest-mock==3.14.0
python-magic==0.4.27
Pillow==10.2.0
pypdf==4.1.0
reportlab==4.1.0